Data.Type.Predicate.Logic

Documentation

type Evident = (TyPred Sing :: Predicate k)

The always-true predicate.

Evident :: Predicate k

data Impossible :: Predicate k

data Not :: Predicate k -> Predicate k

Not p is the predicate that p is not true.

decideNot :: forall p a. Decision (p @@ a) -> Decision (Not p @@ a)

Decide Not p based on decisions of p.

data (&&&) :: Predicate k -> Predicate k -> Predicate k infixr 3

p &&& q is a predicate that both p and q are true.

decideAnd :: forall p q a. Decision (p @@ a) -> Decision (q @@ a) -> Decision ((p &&& q) @@ a)

Decide p &&& q based on decisions of p and q.

data (|||) :: Predicate k -> Predicate k -> Predicate k infixr 2

p ||| q is a predicate that either p and q are true.

decideOr :: forall p q a. Decision (p @@ a) -> Decision (q @@ a) -> Decision ((p ||| q) @@ a)

Decide p ||| q based on decisions of p and q.

type (^^^) p q = (p &&& Not q) ||| (Not p &&& q)

p ^^^ q is a predicate that either p and q are true, but not both.

decideXor :: forall p q a. Decision (p @@ a) -> Decision (q @@ a) -> Decision ((p ^^^ q) @@ a)

Decide p ^^^ q based on decisions of p and q.

data (==>) :: Predicate k -> Predicate k -> Predicate k infixr 2

p ==> q is true if q is provably true under the condition that p is true.

proveImplies :: Prove q -> Prove (p ==> q)

If q is provable, then so is p ==> q.

This can be used as an easy plug-in Provable instance for p ==> q if q is Provable:

instance Provable (p ==> MyPred) where
    prove = proveImplies @MyPred

This instance isn't provided polymorphically because of overlapping instance issues.

explosion :: Impossible --> p

From Impossible a, you can prove anything. Essentially a lifted version of absurd.

atom :: p --> Evident

Evident can be proven from all predicates.

excludedMiddle :: (p &&& Not p) --> Impossible

We cannot have both p and Not p.