Data.Type.Predicate.Param

Documentation

type ParamPred k v = k -> Predicate v

A parameterized predicate.

data Found :: ParamPred k v -> Predicate k

Convert a parameterized predicate into a predicate on the parameter

A Found p is a predicate on p :: ParamPred k v that tests a k for the fact that there exists a v where ParamPred k v is satisfied.

data FlipPP :: ParamPred v k -> ParamPred k v

Flip the arguments of a ParamPred.

data PPMap :: (k ~> j) -> ParamPred j v -> ParamPred k v

Pre-compose a function to a ParamPred. Is essentially flip (.), but unfortunately defunctionalization doesn't work too well with that definition.

class Searchable p where

Minimal complete definition

Nothing

Methods

search :: Decide (Found p)

search :: Findable p => Decide (Found p)

class Searchable p => Findable p where

Methods

find :: Prove (Found p)

data AnyMatch f :: ParamPred k v -> ParamPred (f k) v

AnyMatch f takes a parmaeterized predicate on k (testing for a v) and turns it into a parameterized predicate on f k (testing for a v).

An AnyMatch f p as is a predicate taking an argument a and testing if p a :: Predicate k is satisfied for any item in as :: f k.

A ParamPred k v tests if a k can create some v. The resulting Param (f k) v tests if any k in f k can create some v.