Data.Type.Search

Documentation

type ParamPred k v = k -> Predicate v

A parameterized predicate.

data Found :: ParamPred k v -> Predicate k

Convert a parameterized predicate into a predicate on the parameter

data FlipPP :: ParamPred v k -> ParamPred k v

Flip the arguments of a ParamPred.

class Search (p :: ParamPred k v) where

Methods

search :: Test (Found p)

search :: Search_ p => Test (Found p)

data AnyMatch f :: ParamPred k v -> ParamPred (f k) v

AnyMatch f takes a parmaeterized predicate on k (testing for a v) and turns it into a parameterized predicate on f k (testing for a v).

An AnyMatch f p as is a predicate taking an argument a and testing if p a :: Predicate k is satisfied for any item in as :: f k.