Data.Type.Universe

Documentation

type family Elem (f :: Type -> Type) :: f k -> k -> Type

A witness for membership of a given item in a type-level collection

type In (f :: Type -> Type) (as :: f k) = TyCon1 (Elem f as)

In f as is a predicate that a given input a is a member of collection as.

class Universe (f :: Type -> Type) where

Typeclass for a type-level container that you can quantify or lift type-level predicates over.

Methods

idecideAny

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (Any f p @@ as)
predicate on collection

decideAny, but providing an Elem. See decideAny for more information.

idecideAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (All f p @@ as)
predicate on collection

decideAll, but providing an Elem. See decideAll for more information.

igenAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). Applicative h

=> (forall a. Elem f as a -> Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All f p @@ as)
predicate on collection in context

genAllA, but providing an Elem. See genAllA for more information.

decideAny

Arguments

:: forall (p :: k ~> Type). Universe f

=> Decide p
predicate on value

-> Decide (Any f p)
predicate on collection

You should read this type as:

decideAny' :: (Sing a  -> Decision (p a)    )
           -> (Sing as -> Decision (Any p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if any item in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests existential quantification.

decideAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Decide p
predicate on value

-> Decide (All f p)
predicate on collection

You should read this type as:

decideAll' :: (Sing a  -> Decision (p a)    )
           -> (Sing as -> Decision (All p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if all items in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests universal quantification.

genAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). (Universe f, Applicative h)

=> (forall a. Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All f p @@ as)
predicate on collection in context

If p a is true for all values a in as under some (Applicative) context h, then you can create an All p as under that Applicative context h.

Can be useful with Identity (which is basically unwrapping and wrapping All), or with Maybe (which can express predicates that are either provably true or not provably false).

In practice, this can be used to iterate and traverse and sequence actions over all "items" in as.

genAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Prove p
always-true predicate on value

-> Prove (All f p)
always-true predicate on collection

igenAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). Universe f

=> (forall a. Elem f as a -> Sing a -> p @@ a)
always-true predicate on value

-> Sing as -> All f p @@ as
always-true predicate on collection

If p a is true for all values a in as, then we have All p as. Basically witnesses the definition of All.

foldMapUni :: forall f k (as :: f k) m. (Universe f, Monoid m) => (forall (a :: k). Sing a -> m) -> Sing as -> m

A foldMap over all items in a collection.

ifoldMapUni :: forall f k (as :: f k) m. (Universe f, Monoid m) => (forall a. Elem f as a -> Sing a -> m) -> Sing as -> m

foldMapUni but with access to the index.

index

Arguments

:: Universe f

=> Elem f as a
Witness

-> Sing as
Collection

-> Sing a

Extract the item from the container witnessed by the Elem

pickElem :: forall f k (as :: f k) a. (Universe f, SingI as, SingI a, SDecide k) => Decision (Elem f as a)

Automatically generate a witness for a member, if possible

data All f :: (k ~> Type) -> f k ~> Type

WitAll, but as a predicate. An All f p is a predicate testing a collection as :: f a for the fact that all items in as satisfy p.

This is mostly useful for its Decidable, Provable, and TFunctor instances, which lets you lift predicates on p to predicates on All f p.

newtype WitAll f p (as :: f k)

A WitAll p as is a witness that the predicate p a is true for all items a in the type-level collection as.

Constructors

WitAll

Fields

runWitAll :: forall a. Elem f as a -> p @@ a

data Any f :: (k ~> Type) -> f k ~> Type

WitAny, but as a predicate. An Any f p is a predicate testing a collection as :: f a for the fact that at least one item in as satisfies p.

This is mostly useful for its Decidable and TFunctor instances, which lets you lift predicates on p to predicates on Any f p.

data WitAny f :: (k ~> Type) -> f k -> Type where

A WitAny p as is a witness that, for at least one item a in the type-level collection as, the predicate p a is true.

Constructors

WitAny :: Elem f as a -> (p @@ a) -> WitAny f p as

data Index :: [k] -> k -> Type where

Witness an item in a type-level list by providing its index.

Constructors

IZ :: Index (a ': as) a

IS :: Index bs a -> Index (b ': bs) a

data IsJust :: Maybe k -> k -> Type where

Witness an item in a type-level Maybe by proving the Maybe is Just.

Constructors

IsJust :: IsJust (Just a) a

data IsRight :: Either j k -> k -> Type where

Witness an item in a type-level Either j by proving the Either is Right.

Constructors

IsRight :: IsRight (Right a) a

data NEIndex :: NonEmpty k -> k -> Type where

Witness an item in a type-level NonEmpty by either indicating that it is the "head", or by providing an index in the "tail".

Constructors

NEHead :: NEIndex (a :| as) a

NETail :: Index as a -> NEIndex (b :| as) a

data Snd :: (j, k) -> k -> Type where

Trivially witness an item in the second field of a type-level tuple.

Constructors

Snd :: Snd '(a, b) b

type Null f = (Not (NotNull f) :: Predicate (f k))

type NotNull f = (Any f Evident :: Predicate (f k))

choice :: forall f k (as :: f k). (NotNull f @@ as) -> Σ k (In f as)

From a witness that as is NotNull (non-empty), prove that an item exists in it.

fromChoice :: forall f k (as :: f k) a. SingI a => (In f as @@ a) -> NotNull f @@ as

If an item is in as, prove that as is NotNull (non-empty).

noChoice :: forall f k (as :: f k) a. SingI a => (Null f @@ as) -> Not (In f as) @@ a

If as is Null (empty), prove that any a is not in as.

fromNoChoice :: forall f k (as :: f k). Refuted (Σ k (In f as)) -> Null f @@ as

If it is impossible to produce any item in as, prove that a is Null (empty).