Data.Type.Universe.Subset

Documentation

data Subset f :: (k ~> Type) -> f k ~> Type

newtype WitSubset f p (as :: f k)

A Subset p as describes a decidable subset of type-level collection as.

Constructors

WitSubset

Fields

runWitSubset :: forall a. Elem f as a -> Decision (p @@ a)

makeSubset :: forall f k p (as :: f k). Universe f => (forall a. Elem f as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Subset f p @@ as

Create a Subset from a predicate.

subsetToList :: forall f p t. (Universe f, Alternative t) => (Subset f p --># Any f p) t

Turn a Subset into a list (or any Alternative) of satisfied predicates.

subsetToAny :: forall f p. Universe f => Subset f p -?> Any f p

Restrict a Subset to a single (arbitrary) member, or fail if none exists.

subsetToAll :: forall f p. Universe f => Subset f p -?> All f p

Test if a subset is equal to the entire original collection

intersection :: forall f p q. (Subset f p &&& Subset f q) --> Subset f (p &&& q)

Subset intersection

union :: forall f p q. (Subset f p &&& Subset f q) --> Subset f (p ||| q)

Subset union

symDiff :: forall f p q. (Subset f p &&& Subset f q) --> Subset f (p ^^^ q)

Symmetric subset difference

mergeSubset :: forall f k p q r (as :: f k). (forall a. Decision (p @@ a) -> Decision (q @@ a) -> Decision (r @@ a)) -> (Subset f p @@ as) -> (Subset f q @@ as) -> Subset f r @@ as

Combine two subsets based on a decision function

imergeSubset :: forall f k p q r (as :: f k). (forall a. Elem f as a -> Decision (p @@ a) -> Decision (q @@ a) -> Decision (r @@ a)) -> (Subset f p @@ as) -> (Subset f q @@ as) -> Subset f r @@ as

Combine two subsets based on a decision function