Data.Type.Quantification

Documentation

data Any f :: (k ~> Type) -> f k ~> Type

data WitAny f :: (k ~> Type) -> f k -> Type where

An Any p as is a witness that, for at least one item a in the type-level collection as, the predicate p a is true.

Constructors

WitAny :: Elem f as a -> (p @@ a) -> WitAny f p as

entailAny :: forall f p q. Universe f => (p --> q) -> Any f p --> Any f q

ientailAny

Arguments

:: (Universe f, SingI as)

=> (forall a. Elem f as a -> Sing a -> (p @@ a) -> q @@ a)
implication

-> (Any f p @@ as)

-> Any f q @@ as

If there exists an a s.t. p a, and if p implies q, then there must exist an a s.t. q a.

entailAnyF

Arguments

:: (Universe f, Functor h)

=> (p --># q) h
implication in context

-> (Any f p --># Any f q) h

entailAnyF, but without the membership witness.

ientailAnyF

Arguments

:: Functor h

=> (forall a. Elem f as a -> (p @@ a) -> h (q @@ a))
implication in context

-> (Any f p @@ as)

-> h (Any f q @@ as)

If p implies q under some context h, and if there exists some a such that p a, then there must exist some a such that p q under that context h.

h might be something like, say, Maybe, to give predicate that is either provably true or unprovably false.

Note that it is not possible to do this with p a -> Decision (q a). This is if the p a -> Decision (q a) implication is false, there it doesn't mean that there is no a such that q a, necessarily. There could have been an a where p does not hold, but q does.

data All f :: (k ~> Type) -> f k ~> Type

newtype WitAll f p (as :: f k)

An All p as is a witness that, the predicate p a is true for all items a in the type-level collection as.

Constructors

WitAll

Fields

runWitAll :: forall a. Elem f as a -> p @@ a

entailAll :: forall f p q. Universe f => (p --> q) -> All f p --> All f q

ientailAll

Arguments

:: (Universe f, SingI as)

=> (forall a. Elem f as a -> Sing a -> (p @@ a) -> q @@ a)
implication

-> (All f p @@ as)

-> All f q @@ as

If for all a we have p a, and if p implies q, then for all a we must also have p a.

entailAllF

Arguments

:: (Universe f, Applicative h)

=> (p --># q) h
implication in context

-> (All f p --># All f q) h

entailAllF, but without the membership witness.

ientailAllF

Arguments

:: (Universe f, Applicative h, SingI as)

=> (forall a. Elem f as a -> (p @@ a) -> h (q @@ a))
implication in context

-> (All f p @@ as)

-> h (All f q @@ as)

If p implies q under some context h, and if we have p a for all a, then we must have q a for all a under context h.

decideEntailAll :: forall f p q. Universe f => (p -?> q) -> All f p -?> All f q

decideEntailAll, but without the membeship witness.

idecideEntailAll

Arguments

:: (Universe f, SingI as)

=> (forall a. Elem f as a -> (p @@ a) -> Decision (q @@ a))
decidable implication

-> (All f p @@ as)

-> Decision (All f q @@ as)

If we have p a for all a, and p a can be used to test for q a, then we can test all as for q a.